requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
discord.py==1.7.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
flask-unittest==0.1.1
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
module==0.0.4
multidict==5.1.0
pycparser==2.20
pymongo==3.11.4
python-dotenv==0.15.0
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.3
